pyproject.toml
[project]
name = "zabor"
version = "0.1.0"
description = "zaBor: production-grade crypto trading scripts (research, backtests, execution utilities, analytics)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "tenacity>=8.3",
  "PyYAML>=6.0",
  "pandas>=2.2",
  "numpy>=1.26",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "mypy>=1.10",
  "ruff>=0.6",
  "types-PyYAML",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_any_generics = true

[tool.pytest.ini_options]
testpaths = ["tests"]
